third_party/blink/web_tests/wpt_internal/html/canvas/context-lost/2d-invalid-size-and-gpu-loss.html

<!doctype html>
<meta name="timeout" content="long">
<meta name="variant" content="?testType=html&1-3">
<meta name="variant" content="?testType=html&4-6">
<meta name="variant" content="?testType=html&7-last">
<meta name="variant" content="?testType=html&desynchronized&1-3">
<meta name="variant" content="?testType=html&desynchronized&4-6">
<meta name="variant" content="?testType=html&desynchronized&7-last">
<meta name="variant" content="?testType=detached_html&1-3">
<meta name="variant" content="?testType=detached_html&4-6">
<meta name="variant" content="?testType=detached_html&7-last">
<meta name="variant" content="?testType=offscreen&1-3">
<meta name="variant" content="?testType=offscreen&4-6">
<meta name="variant" content="?testType=offscreen&7-last">
<meta name="variant" content="?testType=offscreen&desynchronized&1-3">
<meta name="variant" content="?testType=offscreen&desynchronized&4-6">
<meta name="variant" content="?testType=offscreen&desynchronized&7-last">
<meta name="variant" content="?testType=placeholder&1-3">
<meta name="variant" content="?testType=placeholder&4-6">
<meta name="variant" content="?testType=placeholder&7-last">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-promise-test.js"></script>
<script src="/common/subset-tests.js"></script>
<script src="./2d-context-lost-utils.js"></script>
<script>
assert_true(!!window.chrome && !!chrome.gpuBenchmarking,
            'This test requires chrome.gpuBenchmarking.');

var desynchronized = false;
(function parseVariants() {
  const params = new URLSearchParams(self.location.search);
  desynchronized = (params.get('desynchronized') !== null);
})();

/**
 * Kill the GPU process and let it restore normally.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas has a provider and kill the GPU process.
  drawAndReadBack(ctx);
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should start with acceleration enabled.');
  terminateGpuProcess(test);
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    await waitForContextLost(ctx);
    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should be restored with GPU acceleration.');
}, 'Valid provider -> GPU crash -> notify GPU lost -> contextlost -> ' +
   'try restore -> restored.',
   // `chrome.gpuBenchmarking` is not available on workers.
   {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Create a resource provider while the GPU process is down.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  terminateGpuProcess(test);

  // Try to create a provider while the GPU process is down.
  if (internals.runtimeFlags.accelerated2dCanvasEnabled) {
    assert_array_equals(drawAndReadBack(ctx), [0, 0, 0, 0],
                        'GPU canvas should be unusable.');
  } else {
    assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                        'CPU canvas should still be usable.');
  }
  assert_false(ctx.isContextLost(), `Canvas shouldn't know the GPU is gone.`);

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    // The provider created while the GPU process was down still gets notified.
    await waitForContextLost(ctx);
    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');

  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should be restored with GPU acceleration.');
}, 'No provider -> GPU crash -> create provider.',
   // `chrome.gpuBenchmarking` is not available on workers.
   {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Resize and reuse a canvas while the GPU is down.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas is initialized with valid resources.
  drawAndReadBack(ctx);
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should start with acceleration enabled.');

  // Kill the GPU process and resize the canvas before it's notified of the GPU
  // process loss.
  terminateGpuProcess(test);
  canvas.width = 64;
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);
  if (internals.runtimeFlags.accelerated2dCanvasEnabled) {
    assert_array_equals(drawAndReadBack(ctx), [0, 0, 0, 0],
                        'GPU canvas should be unusable.');
  } else {
    assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                        'CPU canvas should still be usable.');
  }

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    // The provider created while the GPU process was down still gets notified.
    await waitForContextLost(ctx);
    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should now be usable.');

  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'GPU canvas should be restored with acceleration enabled.');
}, 'Valid provider -> GPU crash -> resize -> create provider.',
    // `chrome.gpuBenchmarking` is not available on workers.
    {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Kill the GPU process and resize the canvas. The canvas can't be notified of
 * the context loss. After the GPU process is restored, canvas can be created
 * as if nothing happened.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas is initialized with valid resources.
  drawAndReadBack(ctx);

  // Kill the GPU process and reset the size before the canvas is notified of
  // the GPU process loss.
  const gpuProcessRestoredPromise = terminateGpuProcess(test);
  canvas.width = 64;

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    // TODO: crbug.com/348445884 - `HTMLCanvasElement` and `OffscreenCanvas`
    // behave differently. This test is merely testing existing behavior, but
    // this should be fixed such that the two classes behave the same way.
    if (HTML_CANVAS_ELEMENT_TEST_TYPES.includes(canvasType)) {
      // Wait for the GPU process to be back. The canvas can't be notified
      // because it doesn't have a provider.
      await resolvedWithoutContextEvent(gpuProcessRestoredPromise, ctx);
      await waitFramesAndRejectOnContextEvents(ctx);

      assert_false(ctx.isContextLost(),
                  `Canvas shouldn't be lost because it has no resources.`);
    } else {
      // `OffscreenCanvas` rasterizes a blank frame after a resize, forcing
      // the creation of a provider while the GPU context is lost.
      await waitForContextLost(ctx);
      await waitForContextRestored(ctx);
    }
  }

  // Using the canvas now recreates a valid provider.
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
}, 'Valid provider -> GPU crash -> resize -> GPU restored -> create provider.',
    // `chrome.gpuBenchmarking` is not available on workers.
    {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Crash the GPU process and give the canvas an invalid size before it's
 * notified of the GPU context loss. The context will only restore after giving
 * the canvas a valid size.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas has a provider and kill the GPU process.
  // Set an invalid size before the canvas is notified of the GPU process loss.
  drawAndReadBack(ctx);
  const gpuProcessRestoredPromise = terminateGpuProcess(test);
  canvas.width = 100000000;
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);

  // Using the canvas now recreates a provider, which fails because of the
  // invalid size.
  drawAndReadBack(ctx);
  await waitForContextLost(ctx);  // Invalid size context lost.

  // Wait for the GPU to be back. The canvas should not restore because the size
  // is still invalid.
  await resolvedWithoutContextEvent(gpuProcessRestoredPromise, ctx);
  await waitFramesAndRejectOnContextEvents(ctx);

  canvas.width = 64;
  await waitForContextRestored(ctx);
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
}, 'Valid provider -> GPU crash -> bad size -> invalid size contextlost -> ' +
   'good size -> contextrestored.',
    // `chrome.gpuBenchmarking` is not available on workers.
    {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Kill the GPU process, cycle the canvas size as good -> bad -> good and
 * recreate a provider before the canvas is notified of the context loss.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas has a provider and kill the GPU process.
  drawAndReadBack(ctx);
  terminateGpuProcess(test);
  canvas.width = 100000000;
  canvas.width = 64;
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);

  drawAndReadBack(ctx);

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    await waitForContextLost(ctx);
    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
}, 'Valid provider -> GPU crash -> bad size -> good size -> create provider.',
    // `chrome.gpuBenchmarking` is not available on workers.
    {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Lose the GPU process and do good size -> bad size -> good size from inside
 * the contextlost event handler.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas has a provider and kill the GPU process.
  drawAndReadBack(ctx);
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should start with acceleration enabled.');
  terminateGpuProcess(test);
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    await waitForContextLost(ctx);

    canvas.width = 100000000;
    canvas.width = 64;

    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should be restored with GPU acceleration.');
}, 'Valid provider -> GPU crash -> contextlost -> bad size -> ' +
   'good size -> try restore scheduled -> try restore -> contextrestored.',
   // `chrome.gpuBenchmarking` is not available on workers.
   {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Lose the GPU process and do good size -> bad size -> good size after the
 * contextlost event handler but before contextrestored.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas has a provider and kill the GPU process.
  drawAndReadBack(ctx);
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should start with acceleration enabled.');
  terminateGpuProcess(test);
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    await waitForContextLost(ctx);
    // We are now inside the contextlost event handler. Wait until the task is
    // finished. A microtask is needed to make sure we regain control before the
    // contextrestored event fires.
    await new Promise(resolve => queueMicrotask(resolve));

    canvas.width = 100000000;
    canvas.width = 64;

    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should be restored with GPU acceleration.');
}, 'Valid provider -> GPU crash -> contextlost -> try restore scheduled -> ' +
   'bad size -> good size -> try restore -> contextrestored.',
   // `chrome.gpuBenchmarking` is not available on workers.
   {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Lose the GPU process and given the canvas an invalid size from inside
 * contextlost event handler.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas has a provider and kill the GPU process.
  drawAndReadBack(ctx);
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should start with acceleration enabled.');
  terminateGpuProcess(test);
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    await waitForContextLost(ctx);

    canvas.width = 100000000;
    // The context tries to restore after 500ms, which will fail because the
    // canvas size is invalid. Wait a whole second to make sure this failure
    // triggers.
    await new Promise(resolve => setTimeout(resolve, 1000));

    // Now restore the size. The canvas should restore on next task.
    canvas.width = 64;
    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should be restored with GPU acceleration.');
}, 'Valid provider -> GPU crash -> contextlost -> bad size -> try restore -> ' +
   'good size -> contextrestored.',
   // `chrome.gpuBenchmarking` is not available on workers.
   {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

/**
 * Lose the GPU process and given the canvas an invalid size after the
 * contextlost event handler ran.
 **/
subsetTest(canvasPromiseTest, async (canvas, {test, canvasType} = {}) => {
  const ctx = get2dContext(canvas, {desynchronized});

  // Make sure the canvas has a provider and kill the GPU process.
  drawAndReadBack(ctx);
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should start with acceleration enabled.');
  terminateGpuProcess(test);
  assert_false(ctx.isContextLost(), `Canvas doesn't yet know the GPU is gone.`);

  if (canvasType === CanvasTestType.OFFSCREEN &&
      !internals.runtimeFlags.accelerated2dCanvasEnabled) {
    // Unaccelerated `OffscreenCanvas` never lose the context.
    await waitFramesAndRejectOnContextEvents(ctx);
  } else {
    await waitForContextLost(ctx);
    // We are now inside the contextlost event handler. Wait until the task is
    // finished. A microtask is needed to make sure we regain control before the
    // contextrestored event fires.
    await new Promise(resolve => queueMicrotask(resolve));

    canvas.width = 100000000;
    // The context tries to restore after 500ms, which will fail because the
    // canvas size is invalid. Wait a whole second to make sure this failure
    // triggers.
    await new Promise(resolve => setTimeout(resolve, 1000));

    // Now restore the size. The canvas should restore on next task.
    canvas.width = 64;
    await waitForContextRestored(ctx);
  }
  assert_array_equals(drawAndReadBack(ctx), [255, 0, 0, 255],
                      'The canvas should be usable.');
  assert_equals(internals.isCanvasImageSourceAccelerated(canvas),
                internals.runtimeFlags.accelerated2dCanvasEnabled,
                'Accelerated canvas should be restored with GPU acceleration.');
}, 'Valid provider -> GPU crash -> contextlost -> try restore scheduled -> ' +
   'bad size -> try restore -> good size -> contextrestored.',
   // `chrome.gpuBenchmarking` is not available on workers.
   {testTypes: MAIN_THREAD_CANVAS_TEST_TYPES});

</script>
